[API] move validation in target profile validation folder
